+++ CHANGELOG.md
@@ -6,14 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixed
-- Removed flicker when changing tabs to "First Order"
+
+## [0.3.3] - 2018-11-01
 
 ### Added
 - Error handling for geolocation
 
 ### Changed
 - Set timeout and high accuracy for geolocation
+
+### Fixed
+- Removed flicker when changing tabs to "First Order"
+- Fix input prefix and suffix breaking with error messages
+
+### Fixed
 
 ## [0.3.2] - 2018-11-01
 ### Fixed